Add files via upload
This code initialises an http server (on 192.168.4.1). Receives POST data which includes WiFi credentials (Home network to connect to), mqtt broker ip and the topic to subscribe to.
Once configured, esp connects to the desired home wifi network, subscribes to the desired topic after connecting to the broker (broker should be on the same wifi network)  and stores all this data on EEPROM. On next power up/reset, esp fetches this data from EEPROM and automatically connects to the wifi network, connects with the broker and subscribes to the same topic again.